refactor: split into macro global enum type
type: move type checker to type

shader: merge material definitions and rename variables

move physics constant

common: cleanup
