+++ Engine/common.hlsli
@@ -1,16 +1,9 @@
-#ifndef __COMMON_HLSL_
-#define MAX_NUM_LIGHTS 8
+#ifndef __COMMON_HLSLI__
 #define TRIANGLE_MACRO 3
-#define MAX_NUM_CASCADES 3
-#define MAX_BONE_COUNT 4
+#include "type.hlsli"
 
 SamplerState           PSSampler : register(s0);
 SamplerComparisonState PSShadowSampler : register(s1);
-
-struct BoneTransformElement
-{
-    matrix transform;
-};
 
 Texture2D                              shaderTexture : register(t0);
 Texture2D                              shaderNormalMap : register(t1);
@@ -22,12 +15,14 @@ static const float4 g_ambientColor = float4(0.15f, 0.15f, 0.15f, 1.0f);
 
 cbuffer PerspectiveBuffer : register(b0)
 {
-    matrix g_cam_world;
-    matrix g_cam_view;
-    matrix g_cam_projection;
-    matrix g_cam_invView;
-    matrix g_cam_invProjection;
-    matrix g_cam_reflectView;
+    matrix g_camWorld;
+    matrix g_camView;
+    matrix g_camProj;
+
+    matrix g_camInvView;
+    matrix g_camInvProj;
+
+    matrix g_camReflectView;
 };
 
 cbuffer TransformBuffer : register(b1)
@@ -43,94 +38,28 @@ cbuffer LightBuffer : register(b2)
     float3 ___p0;
 }
 
-cbuffer SpecularBuffer : register(b3)
-{
-    float  g_specularPower;
-    float3 ___p1;
-    float4 g_specularColor;
-}
-
-struct CascadeShadow
-{
-    matrix g_shadow_view[MAX_NUM_CASCADES];
-    matrix g_shadow_proj[MAX_NUM_CASCADES];
-    float4 g_shadow_z_clip[MAX_NUM_CASCADES];
-};
-
 // current light view and projection matrix of each cascade
-cbuffer CascadeShadowBuffer : register(b4)
+cbuffer CascadeShadowBuffer : register(b3)
 {
     CascadeShadow g_currentShadow;
 }
 
-cbuffer CascadeShadowChunk : register(b5)
+cbuffer CascadeShadowChunk : register(b4)
 {
     CascadeShadow g_cascadeShadowChunk[MAX_NUM_LIGHTS];
 }
 
-cbuffer WaterBuffer : register(b6)
+cbuffer MaterialBuffer : register(b5)
 {
-    float  g_waterTranslation;
-    float  g_reflfrScale;
-    float2 ___p2;
+    float g_specularPower;
+    float g_reflectionTranslation;
+    float g_reflectionScale;
+    float g_refractionScale;
+
+    float4 g_overrideColor;
+    float4 g_specularColor;
+    float4 g_clipPlane;
 }
-
-cbuffer ClipPlaneBuffer : register(b7)
-{
-    float4 g_clip_plane;
-}
-
-struct VertexBoneElement
-{
-    int   boneIndex[MAX_BONE_COUNT] : BONEINDEX;
-    float boneWeight[MAX_BONE_COUNT] : BONEWEIGHT;
-    uint  bone_count : BONECOUNT;
-};
-
-struct VertexInputType
-{
-    float3 position : POSITION;
-    float4 color : COLOR;
-    float2 tex : TEXCOORD0;
-
-    float3 normal : NORMAL;
-    float3 tangent : TANGENT;
-    float3 binormal : BINOARML;
-
-    VertexBoneElement bone_element : BONE;
-};
-
-struct GeometryShadowInputType
-{
-    float4 position : SV_POSITION;
-};
-
-struct PixelShadowStage1InputType
-{
-    float4 position : SV_POSITION;
-    uint   RTIndex : SV_RenderTargetArrayIndex;
-};
-
-struct PixelInputType
-{
-    float4 position : SV_POSITION;
-    float4 world_position : POSITION0;
-    float4 color : COLOR;
-    float2 tex : TEXCOORD0;
-
-    float3 normal : NORMAL;
-    float3 tangent : TANGENT;
-    float3 binormal : BINOARML;
-
-    float4 reflection : POSITION1;
-    float4 refraction : POSITION2;
-
-    float3 viewDirection : TEXCOORD2;
-    float3 lightDirection[MAX_NUM_LIGHTS] : TEXCOORD3;
-
-    float clipSpacePosZ : SV_ClipDistance0;
-    float clipPlane : SV_ClipDistance1;
-};
 
 float4 GetWorldPosition(in matrix mat)
 {
@@ -190,11 +119,11 @@ void GetShadowFactor(
         [unroll] for (j = 0; j < MAX_NUM_CASCADES; ++j)
         {
             const matrix vp = mul(
-                                  g_cascadeShadowChunk[i].g_shadow_view[j],
-                                  g_cascadeShadowChunk[i].g_shadow_proj[j]);
+                                  g_cascadeShadowChunk[i].g_shadowView[j],
+                                  g_cascadeShadowChunk[i].g_shadowProj[j]);
             const float4 position = mul(world_position, vp);
 
-            if (z_clip <= g_cascadeShadowChunk[i].g_shadow_z_clip[j].z)
+            if (z_clip <= g_cascadeShadowChunk[i].g_shadowZClip[j].z)
             {
                 shadowFactor[i] = GetShadowFactorImpl(i, j, position);
                 break;
@@ -208,4 +137,4 @@ float4 LerpShadow(in float4 shadowFactor)
     return lerp(float4(0, 0, 0, 1), float4(1, 1, 1, 1), shadowFactor);
 }
 
-#endif
+#endif // __COMMON_HLSLI__
