+++ Engine/cpp.hint
@@ -16,3 +16,4 @@
 #define INTERNAL_COMP_CHECK_CONSTEXPR(enum_val)
 #define RESOURCE_SELF_INFER_GETTER(TYPE) static inline boost::weak_ptr<TYPE> Get(const std::string& name) { return GetResourceManager().GetResource<TYPE>(name); }
 #define RESOURCE_SELF_INFER_CREATE(TYPE) static inline boost::shared_ptr<TYPE> Create( const std::string& name, const std::filesystem::path& path) { if (const auto check = GetResourceManager(). GetResource<TYPE>(path).lock()) { return check; } const auto obj = boost::make_shared<TYPE>(path); GetResourceManager().AddResource(name, obj); return obj; }
+#define INTERNAL_CB_CHECK_CONSTEXPR(enum_val)
