+++ Engine/gs_cascade_shadow_stage1.hlsl
@@ -2,11 +2,11 @@
 
 [maxvertexcount(TRIANGLE_MACRO * MAX_NUM_CASCADES)] void main(
     triangle GeometryShadowInputType                 input[3],
-    inout TriangleStream<PixelShadowStage1InputType> output)
+    inout TriangleStream<PixelShadowInputType> output)
 {
     for (int i = 0; i < MAX_NUM_CASCADES; ++i)
     {
-        PixelShadowStage1InputType element;
+        PixelShadowInputType element;
         element.RTIndex = i;
 
         for (int j = 0; j < TRIANGLE_MACRO; ++j)
@@ -14,8 +14,8 @@
             element.position =
                     mul(
                         input[j].position, mul(
-                                               g_currentShadow.g_shadow_view[i],
-                                               g_currentShadow.g_shadow_proj[i]));
+                                               g_currentShadow.g_shadowView[i],
+                                               g_currentShadow.g_shadowProj[i]));
             output.Append(element);
         }
